feat(safety-links): embed member panel inline in Household Settings
Extract SafetyLinksPanel component; render inline under Household Settings
instead of a link out. Standalone /safety-links route reuses the same panel
(email deep-links still resolve).
